Restructure files according to popular convention
