+++ lib/active_model/validations/reduce_validator.rb
@@ -1,0 +1,18 @@
+require 'active_model'
+
+module ActiveModel
+  module Validations
+
+    class ReduceValidator < ActiveModel::EachValidator
+      def validate_each(record, attribute, _value)
+        return if record.errors[attribute].size <= 1
+
+        first_message = record.errors[attribute][0]
+
+        record.errors.delete(attribute)
+        record.errors.add(attribute, message: first_message)
+      end
+    end
+
+  end
+end
