+++ lib/reduce_validator/version.rb
@@ -1,0 +1,3 @@
+module ReduceValidator
+  VERSION = '0.1.2'.freeze
+end
