+++ spec/spec_helper.rb
@@ -1,7 +1,6 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'rspec'
 require 'reduce_validator'
-require 'active_model'
 
 class TestModel
   include ActiveModel::Validations
